extract common part of http adapters
