+++ knotx-adapters/knotx-adapter-common/src/test/java/com/cognifide/knotx/adapter/common/http/AllowedHeadersFilterTest.java
@@ -15,10 +15,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.cognifide.knotx.adapter.service.http;
+package com.cognifide.knotx.adapter.common.http;
 
 import com.google.common.collect.Lists;
 
+import com.cognifide.knotx.adapter.common.http.AllowedHeadersFilter;
 import com.cognifide.knotx.adapter.common.http.StringToPatternFunction;
 
 import org.junit.Test;
