+++ knotx-adapters/knotx-adapter-common/src/test/java/com/cognifide/knotx/adapter/common/placeholders/UriTransformerPlaceholderTest.java
